Creating common model file for MNIST
