+++ examples/MNIST/conda.yaml
