--- examples/MNIST/example1/requirements.txt
@@ -1,8 +1,0 @@
-torch==1.6.0
-torchvision==0.7.0
-mlflow==1.11.0
-torchserve
-torch-model-archiver
-matplotlib
-numpy
-
