+++ examples/MNIST/example2/README.md
@@ -16,19 +16,16 @@ Run the following commands to install deployment plugin
 
 ## Generating model file (.pt)
 
-Run the `mnist_model.py` script which will perform training on MNIST handwritten dataset. 
-
-By default,  the script exports the model file as `model_cnn.pt`
+Move to `mlflow-torchserve/examples/MNIST/` directory and run the following command to train the MNIST model  
 
 Command: 
-
 ```
 python mnist_model.py \
     --epochs 5 \
-    --batch-size 64 \
-    --lr 0.01 \
-    --save-model
+    --model-save-path "example2"
 ```
+
+The trained model file `mnist_cnn.pt` and the sample input file `sample.json` will be saved into the model-save-path location(`example2` directory).
 
 ## Starting torchserve
 
@@ -48,7 +45,10 @@ This example demonstrates input as tensor (refer `predict.py` where images are c
 
 Run the following command to create and predict the output based on our test data - `test_data/one.png`
 
-`python predict.py`
+`python predict.py <PATH_TO_MODEL_FILE> <PATH_TO_HANDLER>>`
+
+For ex:
+`python predict.py /home/ubuntu/mlflow-torchserve/examples/MNIST/mnist_model.py mnist_handler.py`
 
 Note: Torchserve plugin determines the version number by itself based on the deployment name. hence, version number 
 is not a mandatory argument for the plugin. For example, by running `predict.py`, a new deployment `mnist_test` is created with version 1.
@@ -57,7 +57,7 @@ If needed, version number can also be explicitly mentioned as a config variable 
 
 ```
 config = {
-    'MODEL_FILE': "mnist_model.py",
+    'MODEL_FILE': "<PATH_TO_MODEL_FILE>",
     'HANDLER': 'mnist_handler.py',
     'VERSION': 5.0
