--- examples/MNIST/example3/MLproject
@@ -1,32 +1,0 @@
-name: mnist-example3
-
-conda_env: conda.yaml
-
-entry_points:
-  main:
-    parameters:
-      epochs: {type: int, default: 5}
-      batch_size: {type: int, default: 64}
-      test_batch_size: {type: int, default: 1000}
-      learning_rate: {type: float, default: 0.01}
-      generate_sample_input: {type: bool, default: True}
-      save_model: {type: bool, default: True}
-      log_interval: {type: int, default: 10}
-      seed: {type: int, default: 1}
-      no_cuda: {type: bool, default: False}
-      gamma: {type: float, default: 0.7}
-      model_save_path: {type: str, default: 'models'}
-
-    command: |
-          python mnist_model.py \
-            --epochs {epochs} \
-            --batch-size {batch_size} \
-            --test-batch-size {test_batch_size} \
-            --lr {learning_rate} \
-            --generate-sample-input {generate_sample_input} \
-            --save-model {save_model} \
-            --log-interval {log_interval} \
-            --seed {seed} \
-            --no-cuda {no_cuda} \
-            --gamma {gamma} \
-            --model-save-path {model_save_path}
