--- examples/MNIST/example3/conda.yaml
@@ -1,14 +1,0 @@
-channels:
-- defaults
-- conda-forge
-- pytorch
-dependencies:
-- python=3.8.2
-- pytorch>=1.6.0
-- torchvision=0.7.0
-- pip
-- pip:
-  - mlflow
-  - cloudpickle==1.6.0
-  - boto3
-  - matplotlib
